+++ src/Monarch/Environment.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * @package Monark
+ * @license http://opensource.org/licenses/MIT
+ */
+
+declare(strict_types=1);
+
+namespace DecodeLabs\Monarch;
+
+interface Environment
+{
+    public string $name { get; }
+    public EnvironmentMode $mode { get; }
+}