+++ CHANGELOG.md
@@ -5,7 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Unreleased
---
+- Swapped Container plugin for active Kingdom instance
+- Moved applicationName to Kingdom interface
+- Added Environment and Build interfaces
+- Removed Container implementation
+- Added Kingdom Service support
+- Removed Veneer dependency
+- Converted to static class implementation
 
 ---